+++ src/main/java/interface_adapter/sign_up/SignUpState.java
@@ -1,4 +1,4 @@
-package interface_adapter;
+package interface_adapter.sign_up;
 
 public class SignUpState {
     private String username = "";
