+++ src/main/java/interface_adapter/sign_up/SignUpViewModel.java
@@ -1,4 +1,6 @@
-package interface_adapter;
+package interface_adapter.sign_up;
+
+import interface_adapter.ViewModel;
 
 import java.beans.PropertyChangeListener;
 import java.beans.PropertyChangeSupport;
