+++ src/main/java/use_case/gcalevent/ExternalEventInputBoundary.java
@@ -1,4 +1,4 @@
-package use_case;
+package use_case.gcalevent;
 
 import java.io.IOException;
