+++ src/main/java/use_case/gcalevent/GCalEventInputData.java
@@ -1,4 +1,4 @@
-package use_case;
+package use_case.gcalevent;
 
 public class GCalEventInputData {
 }
