+++ src/main/java/use_case/local_event/LocalEventInputBoundary.java
@@ -1,4 +1,4 @@
-package use_case;
+package use_case.local_event;
 
 
 public interface LocalEventInputBoundary {
