+++ src/main/java/use_case/user/LogInOutOutputBoundary.java
@@ -1,4 +1,4 @@
-package use_case;
+package use_case.user;
 
 public interface LogInOutOutputBoundary {
     enum ErrorType {USERNAME, PASSWORD};
