+++ src/main/java/use_case/user/UserDataAccessInterface.java
@@ -1,4 +1,4 @@
-package use_case;
+package use_case.user;
 
 import entity.User;
